split Profile out of models
