+++ osidb/dmodels/__init__.py
@@ -1,6 +1,7 @@
+from .profile import Profile
 from .ps_contact import PsContact
 from .ps_module import PsModule
 from .ps_product import PsProduct
 from .ps_update_stream import PsUpdateStream
 
-__all__ = ("PsContact", "PsModule", "PsProduct", "PsUpdateStream")
+__all__ = ("Profile", "PsContact", "PsModule", "PsProduct", "PsUpdateStream")
